+++ include/votca/csg/pairlist.h
@@ -18,11 +18,11 @@
 #ifndef _PAIRLIST_H
 #define	_PAIRLIST_H
 
-#include <list>
 #include <map>
+#include <vector>
 
 namespace votca { namespace csg {
-using namespace std;
+
 
 template<typename element_type, typename pair_type>
 class PairList {
@@ -33,12 +33,12 @@ public:
     //void Generate(BeadList *list1, BeadList *list2 = NULL);    
     void AddPair(pair_type *p);
     
-    typedef typename std::list<pair_type *>::iterator iterator;
+    typedef typename std::vector<pair_type *>::iterator iterator;
     typedef typename std::map<element_type, pair_type*> partners;
     
     iterator begin() { return _pairs.begin(); }
     iterator end() { return _pairs.end(); }
-    typename list<pair_type*>::size_type size() { return _pairs.size(); }    
+    typename std::vector<pair_type*>::size_type size() { return _pairs.size(); }    
     pair_type *front() { return _pairs.front(); }
     pair_type *back() { return _pairs.back(); }    
     bool empty() { return _pairs.empty(); }
@@ -53,9 +53,9 @@ public:
     typedef pair_type pair_t;
 
 protected:
-    list<pair_type *> _pairs;
+    std::vector<pair_type *> _pairs;
       
-    map< element_type , map<element_type, pair_type *> > _pair_map;
+    std::map< element_type , std::map<element_type, pair_type *> > _pair_map;
     
 };
 
@@ -82,7 +82,7 @@ inline void PairList<element_type, pair_type>::Cleanup()
 template<typename element_type, typename pair_type>
 inline pair_type *PairList<element_type, pair_type>::FindPair(element_type e1, element_type e2)
 {
-    typename std::map< element_type , map< element_type, pair_type * > >::iterator iter1;    
+    typename std::map< element_type , std::map< element_type, pair_type * > >::iterator iter1;    
     iter1 = _pair_map.find(e1);
     if(iter1==_pair_map.end()) return NULL;
     
@@ -97,7 +97,7 @@ inline pair_type *PairList<element_type, pair_type>::FindPair(element_type e1, e
 template<typename element_type, typename pair_type>
 typename PairList<element_type, pair_type>::partners *PairList<element_type, pair_type>::FindPartners(element_type e1)
 {
-    typename map< element_type , map<element_type, pair_type *> >::iterator iter;
+    typename std::map< element_type , std::map<element_type, pair_type *> >::iterator iter;
     if((iter=_pair_map.find(e1)) == _pair_map.end())
         return NULL;
     return &(iter->second);