+++ share/scripts/inverse/solve_numpy.sh
@@ -1,6 +1,6 @@
 #! /bin/bash
 #
-# Copyright 2009-2011 The VOTCA Development Team (http://www.votca.org)
+# Copyright 2009-2016 The VOTCA Development Team (http://www.votca.org)
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -20,27 +20,19 @@ cat <<EOF
 ${0##*/}, version %version%
 This script solves a linear equation system from imc using numpy
 
-Usage: ${0##*/} <group> <outfile>
+Usage: ${0##*/} <group> <outfile> <reg> 
 
-Used external packages: numpy
+Uses external packages: numpy
 EOF
    exit 0
 fi
 
-[[ -z $1 || -z $2 ]] && die "${0##*/}: Missing arguments"
+[[ -z $1 || -z $2 || -z $3 ]] && die "${0##*/}: Missing arguments"
 
-# initialize & run the octave file
-cat_external solve numpy | sed -e "s/\$name_out/$2/"  -e "s/\$name/$1/" > solve_$1.sh || die "${0##*/}: sed failed"
-
-#this check is not sufficient for numpy! check for numpy package!
-py="$(csg_get_property cg.inverse.imc.numpy.bin)"
-[ -n "$(type -p $py)" ] || die "${0##*/}: python binary '$py' not found"
-
-critical $py solve_$1.sh
+do_external solve numpy --reg "$3" "$1" "$2"
 
 [[ -f $2 ]] || die "Python failed"
 # temporary compatibility issue
-echo "TODO: check these lines!"
+#TODO: check these lines!
 critical sed -ie 's/NaN/0.0/' "$2"
 critical sed -ie 's/Inf/0.0/' "$2"
-